Rendering framework
Add a Renderer base class which can be used to implement different Renderers.
Add a RenderPipeline class that can render objects it knows Renderers for.
Add a RenderPipelineBuilder class to build RenderPipelines by specifying Renderers.
Add a Compositor class that instructs a RenderPipeline to render a full frame.
Add an EntityRenderer to render entities. This class currently does not render an Entity yet because Entities are not ready to be rendered in current state.
Improvements and cleanup for dependency injection container, game.ts and main.ts files.