Compositor improvements
Initialize the glContext in the Compositor instead of in the RenderPipeline.
Use definitive assignment assertions to mark some of the private variables as initialized. By doing this, the TypeScript compiler no longer shows false errors about the variables not being assigned before their usage.
Add return types to Compositor methods.